aru commit - modularizing js files into sub components
