+++ client/src/components/Events/viewEvents.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {useState,useEffect} from 'react';
 import Axios from 'axios';
-import '../css/eventScreen.css';
+import '../../css/eventScreen.css';
 
 function ViewEvents(props) {
     const [presenter,setpresenter]=useState('');
